Move Hass.io panel files to hassio folder
