+++ hassio/src/addon-store/hassio-addon-repository.js
@@ -2,8 +2,8 @@ import '@polymer/paper-card/paper-card.js';
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import '../../../src/components/hassio-card-content.js';
-import '../../../src/resources/hassio-style.js';
+import '../components/hassio-card-content.js';
+import '../resources/hassio-style.js';
 import NavigateMixin from '../../../src/mixins/navigate-mixin.js';
 
 class HassioAddonRepository extends NavigateMixin(PolymerElement) {
