+++ hassio/src/addon-store/hassio-repositories-editor.js
@@ -5,8 +5,8 @@ import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
 import '../../../src/components/buttons/ha-call-api-button.js';
-import '../../../src/components/hassio-card-content.js';
-import '../../../src/resources/hassio-style.js';
+import '../components/hassio-card-content.js';
+import '../resources/hassio-style.js';
 
 class HassioRepositoriesEditor extends PolymerElement {
   static get template() {
