+++ hassio/src/addon-view/hassio-addon-info.js
@@ -7,9 +7,10 @@ import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
 import '../../../src/components/buttons/ha-call-api-button.js';
 import '../../../src/components/ha-markdown.js';
-import '../../../src/components/hassio-card-content.js';
 import '../../../src/resources/ha-style.js';
 import EventsMixin from '../../../src/mixins/events-mixin.js';
+
+import '../components/hassio-card-content.js';
 
 class HassioAddonInfo extends EventsMixin(PolymerElement) {
   static get template() {
