+++ hassio/src/dashboard/hassio-hass-update.js
@@ -4,8 +4,8 @@ import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
 import '../../../src/components/buttons/ha-call-api-button.js';
-import '../../../src/components/hassio-card-content.js';
-import '../../../src/resources/hassio-style.js';
+import '../components/hassio-card-content.js';
+import '../resources/hassio-style.js';
 
 class HassioHassUpdate extends PolymerElement {
   static get template() {
