+++ hassio/src/snapshots/hassio-snapshots.js
@@ -7,8 +7,8 @@ import '@polymer/paper-radio-group/paper-radio-group.js';
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import '../../../src/components/hassio-card-content.js';
-import '../../../src/resources/hassio-style.js';
+import '../components/hassio-card-content.js';
+import '../resources/hassio-style.js';
 import EventsMixin from '../../../src/mixins/events-mixin.js';
 
 class HassioSnapshots extends EventsMixin(PolymerElement) {
